+++ R/InsertTable.R
@@ -115,11 +115,11 @@ trySettingAutoCommit <- function(connection, value) {
 #' Bulk uploading: 
 #' 
 #' Redshift: The MPP bulk loading relies upon the CloudyR S3 library
-#' to test a connection to an S3 bucket using AWS S3 credentials. Credentials are configured either
+#' to test a connection to an S3 bucket using AWS S3 credentials. Credentials are configured 
 #' directly into the System Environment using the following keys: Sys.setenv("AWS_ACCESS_KEY_ID" =
 #' "some_access_key_id", "AWS_SECRET_ACCESS_KEY" = "some_secret_access_key", "AWS_DEFAULT_REGION" =
 #' "some_aws_region", "AWS_BUCKET_NAME" = "some_bucket_name", "AWS_OBJECT_KEY" = "some_object_key",
-#' "AWS_SSE_TYPE" = "server_side_encryption_type") 
+#' "AWS_SSE_TYPE" = "server_side_encryption_type"). 
 #' 
 #' PDW: The MPP bulk loading relies upon the client
 #' having a Windows OS and the DWLoader exe installed, and the following permissions granted: --Grant
@@ -157,7 +157,7 @@ trySettingAutoCommit <- function(connection, value) {
 #'             dropTableIfExists = TRUE,
 #'             createTable = TRUE,
 #'             tempTable = FALSE,
-#'             bulkLoad = TRUE)  # or, options(databaseConnectorBulkUpload = TRUE)
+#'             bulkLoad = TRUE)  # or, Sys.setenv("DATABASE_CONNECTOR_BULK_UPLOAD" = TRUE)
 #' }
 #' @export
 insertTable <- function(connection,
@@ -168,7 +168,7 @@ insertTable <- function(connection,
                         tempTable = FALSE,
                         oracleTempSchema = NULL,
                         tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
-                        bulkLoad = getOption("databaseConnectorBulkUpload"),
+                        bulkLoad = Sys.getenv("DATABASE_CONNECTOR_BULK_UPLOAD"),
                         useMppBulkLoad = Sys.getenv("USE_MPP_BULK_LOAD"),
                         progressBar = FALSE,
                         camelCaseToSnakeCase = FALSE) {
@@ -184,7 +184,7 @@ insertTable.default <- function(connection,
                                 tempTable = FALSE,
                                 oracleTempSchema = NULL,
                                 tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
-                                bulkLoad = getOption("databaseConnectorBulkUpload"),
+                                bulkLoad = Sys.getenv("DATABASE_CONNECTOR_BULK_UPLOAD"),
                                 useMppBulkLoad = Sys.getenv("USE_MPP_BULK_LOAD"),
                                 progressBar = FALSE,
                                 camelCaseToSnakeCase = FALSE) {
@@ -194,14 +194,12 @@ insertTable.default <- function(connection,
          .frequency_id = "useMppBulkLoad")
     bulkLoad <- useMppBulkLoad
   }
+  bulkLoad <- (!is.null(bulkLoad) && bulkLoad == "TRUE")
   if (!is.null(oracleTempSchema) && oracleTempSchema != "") {
     warn("The 'oracleTempSchema' argument is deprecated. Use 'tempEmulationSchema' instead.",
          .frequency = "regularly",
          .frequency_id = "oracleTempSchema")
     tempEmulationSchema <- oracleTempSchema
-  }
-  if (is.null(bulkLoad) || bulkLoad == "") {
-    bulkLoad = FALSE
   }
   if (camelCaseToSnakeCase) {
     colnames(data) <- SqlRender::camelCaseToSnakeCase(colnames(data))
@@ -357,7 +355,7 @@ insertTable.DatabaseConnectorDbiConnection <- function(connection,
                                                        tempTable = FALSE,
                                                        oracleTempSchema = NULL,
                                                        tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
-                                                       bulkLoad = getOption("databaseConnectorBulkUpload"),
+                                                       bulkLoad = Sys.getenv("DATABASE_CONNECTOR_BULK_UPLOAD"),
                                                        useMppBulkLoad = Sys.getenv("USE_MPP_BULK_LOAD"),
                                                        progressBar = FALSE,
                                                        camelCaseToSnakeCase = FALSE) {